llama-index-weaviate/requirements.txt
weaviate-client
llama-index
sentence-transformers
text_generation